fixing imports and refactoring utils
